streamlit/requirements.txt
streamlit==1.40.1
requests==2.32.3
Pillow==10.3.0